Add 'Satoshi-Light' font to global styles
The 'Satoshi-Light' font is now incorporated into the global styles file as a primary font for the body. It defines the font with different types such as 'woff2', 'woff', and 'truetype' to ensure compatibility across various browsers. Furthermore, the 'swap' setting has been enabled for font-display property to improve webpage load times.