Add loading state to download button
Added a loading state to the download button to improve user experience. This change enables the button to show 'Loading...' text while a download is in progress, making it clear to the user that their action is being processed. Also, the Navbar component was imported and used in place of a static image.