Implement search feature and add Navbar component
Implemented a search feature which supports both YouTube and SoundCloud. This utilizes the 'youtube-sr' and 'soundcloud-scraper' packages to fetch search results from both platforms and return a combined list of results. Additionally, introduced a Navbar component to enhance site navigation. This update added the 'framer-motion' package as a project dependency.