Improve search and download user interfaces.
This update overhauls the user interfaces for search and download functionalities. The changes include implementing a search bar with an integrated button in the search page and adding icons to buttons for better visual interaction. Significant alterations are also done to the Navbar component, giving it a more minimalistic and sleek design. The homepage is simplified and now serves as a navigation guide, leading users to either the search or download page. Finally, the previously combined search and download page is separated into two, enhancing clarity and usability.